Created Canvas to gantry transform and integrated with everything else
-Can now paint any image on any size canvas automagiaclly

+++ HighLevel/decomposition/decomp_color/decomp_color.py
@@ -49,16 +49,33 @@ def color_quantize(image, n_colors=2, size_to=200):
     if size_to>0:
         small_img = util.resize(image,size_to)
 
-    pixel = np.reshape(np.float32(small_img),(small_img.shape[0]*small_img.shape[1],3))
+    # small_img = cv2.cvtColor(np.float32(small_img)*1./255, cv2.COLOR_BGR2LAB)
+
+    pixel = np.reshape(small_img,(small_img.shape[0]*small_img.shape[1],3))
 
     # performing the clustering
-    centroids,_ = kmeans(pixel,n_colors,iter=100)
+    centroids,_ = kmeans(np.float32(pixel),n_colors,iter=100)
+
+    # out  = np.uint8(np.array([centroids]))
+    # cv2.cvtColor(np.array([centroids]), cv2.COLOR_LAB2BGR,out)
+    # centroids = out[0]
+
+    # L = centroids[:,:,0]*100/255
+    # a = centroids[:,:,1]+128
+    # b = centroids[:,:,2]+128
+
+    # centroids = np.uint8(cv2.merge((L,a,b)))
 
     return np.uint8(centroids)
 
 def classify(img_colors, paint_colors):
     # Map old img_colors to their closest color in paint_colors
+    
     colors = np.array(paint_colors)
+
+    # img_colors = cv2.cvtColor(np.array([img_colors],np.uint8), cv2.COLOR_BGR2LAB)[0]
+    # paint_colors = cv2.cvtColor(np.array([paint_colors],np.uint8), cv2.COLOR_BGR2LAB)[0]
+
     qnt,_ = vq(img_colors,paint_colors)
 
     # return list of new paint colors plus canvas color
